Revert "Hopefully, Ram usage reduced"
This reverts commit 78ef32326a38217525ec849d789a60deb8aa93c6.

+++ code/touch/touch_src/touch_coordinate.h
@@ -17,7 +17,7 @@
 typedef struct touch_coordinate_t
 {
 	uint16_t xPos;
-	uint8_t yPos;
+	uint16_t yPos;
 }
 touch_coordinate_t;
